+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/BookingDetailDto.java
@@ -6,13 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookingDetailDto {
   private Long bookingId;
-  private float orderTotal;
+  private BigDecimal orderTotal;
   private String boughtByEmail;
   private BookingType type;
 }
